final bugfixing (hopefully)
