+++ server/src/main/java/com/amos/server/ServerConnectionConnAuth.java
@@ -1,15 +1,8 @@
 package com.amos.server;
 
-import android.content.Intent;
-import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
-import android.util.Log;
-import android.view.KeyEvent;
 import android.view.inputmethod.EditorInfo;
 import android.widget.EditText;
-import android.widget.TextView;
-
-import java.security.SecureRandom;
 
 /**
  * Server-Class soll tun:
@@ -40,23 +33,5 @@ public class ServerConnectionConnAuth extends ServerConnAuthActivity {
             }
             return false;
         });
-    }
-
-    /**
-     * Generates a name for the server.
-     * @return The server name, consisting of R.string.flyinn_server_name + a random string
-     */
-    @Override
-    protected String generateName() {
-        int suffix = 5;
-        String AB = "0123456789ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz";
-        SecureRandom rnd = new SecureRandom();
-
-        StringBuilder sb = new StringBuilder(suffix);
-        for (int i = 0; i < suffix; i++) {
-            sb.append(AB.charAt(rnd.nextInt(AB.length())));
-        }
-
-        return R.string.flyinn_server_name + sb.toString();
     }
 }
